+++ baselines/flwr_baselines/publications/fedprox_mnist/model.py
@@ -1,4 +1,4 @@
-"""CNN model architecutre, training and testing functions for MNIST."""
+"""CNN model architecutre, training, and testing functions for MNIST."""
 
 
 from typing import List, Tuple
@@ -11,8 +11,8 @@ from torch.utils.data import DataLoader
 
 
 class Net(nn.Module):
-    """Convolutional Neural Network architecture as described in McMahan 2017
-    paper :
+    """
+    Convolutional Neural Network architecture as described in McMahan 2017 paper :
 
     [Communication-Efficient Learning of Deep Networks from
     Decentralized Data] (https://arxiv.org/pdf/1602.05629.pdf)