+++ baselines/flwr_baselines/publications/fedprox_mnist/tests/data_test.py
@@ -1,20 +1,7 @@
 """Used to test the model and the data partitionning."""
 
 
-from flwr_baselines.publications.fedprox_mnist import dataset, model
-
-
-def test_cnn_size_mnist() -> None:
-    """Test number of parameters with MNIST-sized inputs."""
-    # Prepare
-    net = model.Net()
-    expected = 1_663_370
-
-    # Execute
-    actual = sum([p.numel() for p in net.parameters()])
-
-    # Assert
-    assert actual == expected
+from flwr_baselines.publications.fedprox_mnist import dataset
 
 
 def test_non_iid_partitionning(num_clients: int = 100) -> None: